pyproject.toml
[tool.poetry]
name = "django-2-dekux-x"
version = "0.1.0"
description = ""
authors = ["Dias <saidshabekov@gmail.com>"]
readme = "README.md"
packages = [{include = "django_2_dekux_x"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.4"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
